ansiblenew

+++ Jenkinsfile
@@ -7,11 +7,17 @@ pipeline {
             }
           }
     stage('Installing ansible') {
-      steps {
+     script {     
+       def ansible_exists = fileExists /usr/bin/ansible
+       if (ansible_exists == true) {
+           echo "Skikking ansible install already exists"
+        } else {
         sh 'sudo apt-get update -y && sudo apt-get upgrade -y'
         sh 'sudo apt install -y wget tree unzip ansible python3-pip python3-apt'
             }
          }
+     }
+ }
     stage('Third stage') {
       steps {
         echo "Third stage"
